chore: prepare for initial release
